Moving state from Sketch to SketchContext
